Renaming unit-test files according to the pattern: 'SomethingTests'
