+++ sansa-query-flink/src/test/scala/net/sansa_stack/query/flink/sparqlify/RdfPartitionFlinkTests.scala
