+++ sansa-query-spark/src/test/scala/net/sansa_stack/query/spark/graph/SparqlToSparkForRDDTests.scala
